Create a base trait for row level operations - Foundational work for MERGE INTO
